fix(inference): improve overload selection
