+++ crates/biome_js_analyze/tests/specs/nursery/noMisusedPromises/checksVoidReturnValid.ts
@@ -20,3 +20,14 @@ new Promise((resolve, reject) => {
     resolve();
   })();
 });
+
+declare function consume(kind: "async", callback: () => Promise<void>): void;
+declare function consume(kind: "sync", callback: () => void): void;
+consume("async", async () => {});
+
+class ConstructorConsumer {
+  constructor(callback: () => void);
+  constructor(callback: () => Promise<void>, marker: number);
+  constructor(_callback: (() => void) | (() => Promise<void>), _marker?: number) {}
+}
+new ConstructorConsumer(async () => {}, 1);
